+++ SplatTagDatabase/Merger.cs
@@ -4,7 +4,6 @@ using System.Collections.Concurrent;
 using System.Collections.Generic;
 using System.Diagnostics;
 using System.Linq;
-using System.Runtime.CompilerServices;
 using System.Threading.Tasks;
 
 namespace SplatTagDatabase
@@ -13,16 +12,20 @@ namespace SplatTagDatabase
   {
     public static void FinalisePlayers(IDictionary<uint, Player> playersToMutate)
     {
-      Debug.WriteLine("Beginning FinalisePlayers.");
       if (playersToMutate == null) return;
 
       var keys = playersToMutate.Keys.Reverse().ToArray();
+      Trace.WriteLine($"Beginning FinalisePlayers on {keys.Length} entries.");
       foreach (var playerId in keys)
       {
         var playersMinusCurrent = playersToMutate.Values.Where(p => p.Id != playerId);
         Debug.Assert(playersMinusCurrent.Count() == playersToMutate.Count - 1);
         var testPlayer = playersToMutate[playerId];
-        Player foundPlayer = FindSamePlayer(playersMinusCurrent, testPlayer, null);
+
+        // Find an equivalent player.
+        Player foundPlayer = FindSamePlayerPersistent(playersMinusCurrent, testPlayer)
+          ?? FindSamePlayerByTeamAndName(playersMinusCurrent, testPlayer);
+
         if (foundPlayer != null)
         {
           Debug.Assert(foundPlayer != testPlayer, $"Player to merge is the same as the one being merged: {foundPlayer.Name} (Id {foundPlayer.Id}) equiv to {testPlayer.Name} (Id [{playerId}] == {testPlayer.Id})");
@@ -30,17 +33,19 @@ namespace SplatTagDatabase
           playersToMutate.Remove(playerId);
         }
       }
+      Trace.WriteLine($"Finished FinalisePlayers with {playersToMutate.Count} entries.");
     }
 
-    public static void CorrectPlayerIds(IEnumerable<Player> incomingPlayers, IDictionary<long, long> teamsMergeResult)
+    public static void CorrectPlayerIds(ICollection<Player> incomingPlayers, IDictionary<long, long> teamsMergeResult)
     {
-      if (incomingPlayers == null || teamsMergeResult == null) return;
-      Debug.WriteLine("Beginning CorrectPlayerIds. Entries: ");
+      if (incomingPlayers == null || teamsMergeResult == null || incomingPlayers.Count == 0 || teamsMergeResult.Count == 0) return;
+
+      Debug.WriteLine($"CorrectPlayerIds called with {incomingPlayers.Count} entries.");
+      Debug.WriteLine("Entries: ");
       foreach (var resultPair in teamsMergeResult)
       {
         Debug.WriteLine($"[{resultPair.Key}] --> {resultPair.Value}");
       }
-
 
       Parallel.ForEach(incomingPlayers, (importPlayer) =>
       {
@@ -80,7 +85,10 @@ namespace SplatTagDatabase
       {
         try
         {
-          Player foundPlayer = FindSamePlayer(playersToMutate.Values, importPlayer, transformedPlayerNames);
+          // First, try match through persistent information only.
+          // If that doesn't work, try and match a name and same team.
+          Player foundPlayer = FindSamePlayerPersistent(playersToMutate.Values, importPlayer)
+            ?? FindSamePlayerByTeamAndName(playersToMutate.Values, importPlayer);
 
           if (foundPlayer == null)
           {
@@ -112,25 +120,28 @@ namespace SplatTagDatabase
     }
 
     /// <summary>
-    /// Find a player that matches another instance.
+    /// Find a player that matches another instance through their persistent information.
     /// </summary>
     /// <param name="playersToMutate">The players to search</param>
     /// <param name="testPlayer">The player instance to try and find</param>
     /// <returns>The matched player, or null if new</returns>
-    private static Player FindSamePlayer(IEnumerable<Player> playersToMutate, Player testPlayer, IEnumerable<KeyValuePair<Player, string[]>> transformedPlayerNames)
+    private static Player FindSamePlayerPersistent(IEnumerable<Player> playersToMutate, Player testPlayer)
     {
       bool tryMatchFcs = !string.IsNullOrEmpty(testPlayer.FriendCode);
       bool tryMatchDiscord = !string.IsNullOrEmpty(testPlayer.DiscordName);
       bool tryMatchDiscordId = testPlayer.DiscordId != null;
+      bool tryMatchTwitch = testPlayer.Twitch != null;
+      bool tryMatchTwitter = testPlayer.Twitter != null;
+      bool tryMatchBattlefySlug = testPlayer.BattlefySlugs?.Count > 0;
+      bool tryMatchBattlefyUsername = testPlayer.BattlefyUsername != null;
 
-      // Quick test all FCs and Discord names.
-      Player foundSame = playersToMutate.FirstOrDefault(p =>
+      return playersToMutate.FirstOrDefault(p =>
       {
         // Test if the Switch FC's match.
         if (tryMatchFcs && p.FriendCode?.Equals(testPlayer.FriendCode) == true)
         {
           // They do.
-          Debug.WriteLine($"FindSamePlayer: Matched player {testPlayer} (Id {testPlayer.Id}) with FC {p.FriendCode} from player {p} (Id {p.Id}).");
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with FC {p.FriendCode} from player {p} (Id {p.Id}).");
           return true;
         }
 
@@ -138,22 +149,96 @@ namespace SplatTagDatabase
         if (tryMatchDiscordId && p.DiscordId?.Equals(testPlayer.DiscordId) == true)
         {
           // They do.
-          Debug.WriteLine($"FindSamePlayer: Matched player {testPlayer} (Id {testPlayer.Id}) with Discord Id {p.DiscordId} from player {p} (Id {p.Id}).");
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with Discord Id {p.DiscordId} from player {p} (Id {p.Id}).");
           return true;
         }
 
         // Test if the Discord names match.
-        if (tryMatchDiscord && p.DiscordName?.Equals(testPlayer.DiscordName, StringComparison.OrdinalIgnoreCase) == true)
+        if (tryMatchDiscord && p.DiscordName?.Equals(testPlayer.DiscordName) == true)
         {
           // They do.
-          Debug.WriteLine($"FindSamePlayer: Matched player {testPlayer} (Id {testPlayer.Id}) with Discord name {p.DiscordName} from player {p} (Id {p.Id}).");
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with Discord name {p.DiscordName} from player {p} (Id {p.Id}).");
+          return true;
+        }
+
+        // Test if the Twitches match.
+        if (tryMatchTwitch && p.Twitch?.Equals(testPlayer.Twitch) == true)
+        {
+          // They do.
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with Twitch {p.Twitch} from player {p} (Id {p.Id}).");
+          return true;
+        }
+
+        // Test if the Twitters match.
+        if (tryMatchTwitter && p.Twitter?.Equals(testPlayer.Twitter) == true)
+        {
+          // They do.
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with Twitter {p.Twitter} from player {p} (Id {p.Id}).");
+          return true;
+        }
+
+        // Test if the Battlefy Usernames match.
+        if (tryMatchBattlefyUsername && p.BattlefyUsername?.Equals(testPlayer.BattlefyUsername) == true)
+        {
+          // They do.
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with BattlefyUsername {p.BattlefyUsername} from player {p} (Id {p.Id}).");
+          return true;
+        }
+
+        // Test if any of the Battlefy Slugs match.
+        if (tryMatchBattlefySlug && p.BattlefySlugs.Intersect(testPlayer.BattlefySlugs).Any())
+        {
+          // They do.
+          Debug.WriteLine($"FindSamePlayerPersistent: Matched player {testPlayer} (Id {testPlayer.Id}) with test BattlefySlug(s) e.g. {testPlayer.BattlefySlugs.First()} with player {p} (Id {p.Id}).");
           return true;
         }
 
         return false;
       });
-      if (foundSame != null) return foundSame;
+    }
 
+    /// <summary>
+    /// Find a player that matches another instance by matching a team and name.
+    /// Does not search other information.
+    /// </summary>
+    /// <param name="playersToMutate">The players to search</param>
+    /// <param name="testPlayer">The player instance to try and find</param>
+    /// <returns>The matched player, or null if new</returns>
+    private static Player FindSamePlayerByTeamAndName(IEnumerable<Player> playersToMutate, Player testPlayer)
+    {
+      // Long test for names with transformation.
+      HashSet<string> testPlayerTransformedNames = new HashSet<string>(testPlayer.Names.Select(n => n.Replace(" ", "").TransformString().ToLowerInvariant()));
+
+      return playersToMutate.FirstOrDefault(p =>
+      {
+        if (p.Teams.Intersect(testPlayer.Teams).Any())
+        {
+          // Replace spaces because people adding tags or starting with space messes up same-name detection.
+          // Test if the name matches the names that we know this player by.
+          foreach (string knownName in p.Names.Select(n => n.Replace(" ", "").TransformString().ToLowerInvariant()))
+          {
+            if ((!string.IsNullOrWhiteSpace(knownName)) && (testPlayerTransformedNames.Contains(knownName)))
+            {
+              Debug.WriteLine($"FindSamePlayerThroughNames: Matched uncached player {testPlayer} (Id {testPlayer.Id}) with known name {knownName} from player {p} (Id {p.Id}).");
+              return true;
+            }
+          }
+        }
+
+        return false;
+      });
+    }
+
+    /// <summary>
+    /// Find a player that matches another instance by any of their known names.
+    /// Does not search other information. Use <see cref="FindSamePlayerPersistent(IEnumerable{Player}, Player)"/> if you want this.
+    /// </summary>
+    /// <param name="playersToMutate">The players to search</param>
+    /// <param name="testPlayer">The player instance to try and find</param>
+    /// <param name="transformedPlayerNames">Cached player names. Can be null to calculate.</param>
+    /// <returns>The matched player, or null if new</returns>
+    private static Player FindSamePlayerByNames(IEnumerable<Player> playersToMutate, Player testPlayer, IEnumerable<KeyValuePair<Player, string[]>> transformedPlayerNames)
+    {
       // Long test for names with transformation.
       HashSet<string> testPlayerTransformedNames = new HashSet<string>(testPlayer.Names.Select(n => n.Replace(" ", "").TransformString().ToLowerInvariant()));
 
@@ -168,7 +253,7 @@ namespace SplatTagDatabase
           {
             if ((!string.IsNullOrWhiteSpace(knownName)) && (testPlayerTransformedNames.Contains(knownName)))
             {
-              Debug.WriteLine($"FindSamePlayer: Matched uncached player {testPlayer} (Id {testPlayer.Id}) with known name {knownName} from player {p} (Id {p.Id}).");
+              Debug.WriteLine($"FindSamePlayerThroughNames: Matched uncached player {testPlayer} (Id {testPlayer.Id}) with known name {knownName} from player {p} (Id {p.Id}).");
               return true;
             }
           }
@@ -185,7 +270,7 @@ namespace SplatTagDatabase
           {
             if ((!string.IsNullOrWhiteSpace(knownName)) && (testPlayerTransformedNames.Contains(knownName)))
             {
-              Debug.WriteLine($"FindSamePlayer: Matched cached player {testPlayer} (Id {testPlayer.Id}) with known name {knownName} from player {tPair.Key} (Id {tPair.Key.Id}).");
+              Debug.WriteLine($"FindSamePlayerThroughNames: Matched cached player {testPlayer} (Id {testPlayer.Id}) with known name {knownName} from player {tPair.Key} (Id {tPair.Key.Id}).");
               return true;
             }
           }